+++ monai/utils/__init__.py
@@ -39,8 +39,6 @@ from .misc import (
     MAX_SEED,
     ImageMetaKey,
     copy_to_device,
-    dtype_numpy_to_torch,
-    dtype_torch_to_numpy,
     ensure_tuple,
     ensure_tuple_rep,
     ensure_tuple_size,
@@ -75,3 +73,10 @@ from .module import (
 )
 from .profiling import PerfContext, torch_profiler_full, torch_profiler_time_cpu_gpu, torch_profiler_time_end_to_end
 from .state_cacher import StateCacher
+from .type_conversion import (
+    convert_data_type,
+    dtype_numpy_to_torch,
+    dtype_torch_to_numpy,
+    get_dtype,
+    get_equivalent_dtype,
+)
